+++ ui-plugins/muse-manager/src/index.js
@@ -1,13 +1,13 @@
 import plugin from 'js-plugin';
 import * as ext from './ext';
 import route from './route';
-// import reducer from './rootReducer';
 import * as hooks from './hooks';
 import * as utils from './utils';
 import * as pm from './features/pm';
 import * as common from './features/common';
 import JavascriptTimeAgo from 'javascript-time-ago';
 import en from 'javascript-time-ago/locale/en';
+
 import './modals';
 import './style.less';
 import { useAbility } from './hooks';
@@ -18,6 +18,7 @@ const InitAbilityComp = () => {
   Object.assign(ability, useAbility());
   return null;
 };
+
 // Initialize the desired locales.
 JavascriptTimeAgo.locale(en);
 plugin.register({