+++ workspace/packages/muse-cra-patch/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ebay/muse-cra-patch
 
+## 1.0.3
+
+### Patch Changes
+
+- Upvers"
+
 ## 1.0.1
 
 ### Patch Changes