+++ workspace/packages/muse-express-middleware/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @ebay/muse-express-middleware
 
+## 1.0.48
+
+### Patch Changes
+
+- Auto detect plugin entry
+
+## 1.0.47
+
+### Patch Changes
+
+- Vite 5.
+
+## 1.0.46
+
+### Patch Changes
+
+- Upvers"
+- Updated dependencies
+  - @ebay/muse-core@1.0.37
+  - @ebay/muse-modules-analyzer@1.0.10
+
+## 1.0.45
+
+### Patch Changes
+
+- vite update
+
 ## 1.0.44
 
 ### Patch Changes