+++ workspace/packages/muse-modules/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ebay/muse-modules
 
+## 1.0.27
+
+### Patch Changes
+
+- Upvers"
+
+## 1.0.26
+
+### Patch Changes
+
+- vite update
+
 ## 1.0.25
 
 ### Patch Changes