+++ workspace/packages/muse-setup-cra/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ebay/muse-setup-cra
 
+## 1.1.2
+
+### Patch Changes
+
+- Upvers"
+
 ## 1.1.1
 
 ### Patch Changes